@@ -11,23 +11,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.0.7",
-    "@angular/common": "^9.0.7",
-    "@angular/compiler": "^9.0.7",
-    "@angular/core": "^9.0.7",
-    "@angular/forms": "^9.0.7",
-    "@angular/localize": "^9.0.7",
-    "@angular/platform-browser": "^9.0.7",
-    "@angular/platform-browser-dynamic": "^9.0.7",
-    "@angular/router": "^9.0.7",
+    "@angular/animations": "^9.1.12",
+    "@angular/common": "^9.1.12",
+    "@angular/compiler": "^9.1.12",
+    "@angular/core": "^9.1.12",
+    "@angular/forms": "^9.1.12",
+    "@angular/localize": "^9.1.12",
+    "@angular/platform-browser": "^9.1.12",
+    "@angular/platform-browser-dynamic": "^9.1.12",
+    "@angular/router": "^9.1.12",
     "@ckeditor/ckeditor5-build-classic": "^18.0.0",
     "@fortawesome/angular-fontawesome": "^0.6.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
     "@ng-bootstrap/ng-bootstrap": "^6.0.2",
+    "@ng-select/ng-select": "^4.0.4",
     "@sentry/browser": "^5.15.0",
-    "@worldskills/bootstrap": "^4.4.0",
-    "@worldskills/worldskills-angular-lib": "^8.0.1",
+    "@worldskills/bootstrap": "^4.4.1",
+    "@worldskills/worldskills-angular-lib": "^9.0.1",
     "angular-oauth2-oidc": "^9.1.0",
     "jquery": "^3.4.1",
     "json2typescript": "^1.2.5",
@@ -40,8 +41,8 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.900.7",
     "@angular/cli": "^9.0.7",
-    "@angular/compiler-cli": "^9.0.7",
-    "@angular/language-service": "^9.0.7",
+    "@angular/compiler-cli": "^9.1.12",
+    "@angular/language-service": "^9.1.12",
     "@types/faker": "^4.1.11",
     "@types/jasmine": "^3.5.10",
     "@types/jasminewd2": "~2.0.3",
