@@ -1,22 +1,31 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 import {QuizzesComponent} from './quizzes/quizzes.component';
-import {QuizzesQuizComponent} from './quizzes/quizzes-quiz/quizzes-quiz.component';
-import {QuizzesQuizQuestionsComponent} from './quizzes/quizzes-quiz/quizzes-quiz-questions/quizzes-quiz-questions.component';
-import {QuizzesQuestionUpdateComponent} from './quizzes/quizzes-question-update/quizzes-question-update.component';
-import {QuizzesQuizPreviewComponent} from './quizzes/quizzes-quiz/quizzes-quiz-preview/quizzes-quiz-preview.component';
-import {QuizzesQuizTranslationsComponent} from './quizzes/quizzes-quiz/quizzes-quiz-translations/quizzes-quiz-translations.component';
-import {QuizzesTranslationUpdateComponent} from './quizzes/quizzes-translation-update/quizzes-translation-update.component';
-import {QuizzesQuizAttemptsComponent} from './quizzes/quizzes-quiz/quizzes-quiz-attempts/quizzes-quiz-attempts.component';
-import {QuizzesAttemptComponent} from './quizzes/quizzes-attempt/quizzes-attempt.component';
+import {QuizzesQuizComponent} from './quizzes-quiz/quizzes-quiz.component';
+import {QuizzesQuizQuestionsComponent} from './quizzes-quiz-questions/quizzes-quiz-questions.component';
+import {QuizzesQuestionUpdateComponent} from './quizzes-question-update/quizzes-question-update.component';
+import {QuizzesQuizPreviewComponent} from './quizzes-quiz-preview/quizzes-quiz-preview.component';
+import {QuizzesQuizTranslationsComponent} from './quizzes-quiz-translations/quizzes-quiz-translations.component';
+import {QuizzesTranslationUpdateComponent} from './quizzes-translation-update/quizzes-translation-update.component';
+import {QuizzesQuizAttemptsComponent} from './quizzes-quiz-attempts/quizzes-quiz-attempts.component';
+import {QuizzesAttemptComponent} from './quizzes-attempt/quizzes-attempt.component';
 import {QuizComponent} from './quiz/quiz.component';
-import {QuizzesQuizUpdateComponent} from './quizzes/quizzes-quiz-update/quizzes-quiz-update.component';
-import {QuizzesQuizCreateComponent} from './quizzes/quizzes-quiz-create/quizzes-quiz-create.component';
-import {QuizzesQuestionCreateComponent} from './quizzes/quizzes-question-create/quizzes-question-create.component';
-import {QuizzesTranslationCreateComponent} from './quizzes/quizzes-translation-create/quizzes-translation-create.component';
+import {QuizzesQuizUpdateComponent} from './quizzes-quiz-update/quizzes-quiz-update.component';
+import {QuizzesQuizCreateComponent} from './quizzes-quiz-create/quizzes-quiz-create.component';
+import {QuizzesQuestionCreateComponent} from './quizzes-question-create/quizzes-question-create.component';
+import {QuizzesTranslationCreateComponent} from './quizzes-translation-create/quizzes-translation-create.component';
 import {HomeComponent} from './home/home.component';
 import {AppAuthGuard} from '@worldskills/worldskills-angular-lib';
 import {ErrorComponent} from './error/error.component';
+
+const APP_CODE = 1300;
+
+function forAppCode(appCode: number, roles: Array<string>) {
+  return roles.map(name => ({
+    appCode,
+    name
+  }));
+}
 
 const routes: Routes = [
   {
@@ -28,27 +37,27 @@ const routes: Routes = [
     pathMatch: 'full',
     component: QuizComponent,
     canActivate: [AppAuthGuard],
-    data: {roles: ['AttemptQuizzes']}
+    data: {roles: forAppCode(APP_CODE, ['AttemptQuizzes'])}
   },
   {
     path: 'quizzes',
     pathMatch: 'full',
     canActivate: [AppAuthGuard],
     component: QuizzesComponent,
-    data: {breadcrumb: 'Quizzes', roles: ['Admin', 'EditQuizzes']}
+    data: {breadcrumb: 'Quizzes', roles: forAppCode(APP_CODE, ['Admin', 'EditQuizzes'])}
   },
   {
     path: 'quizzes/create',
     pathMatch: 'full',
     component: QuizzesQuizCreateComponent,
     canActivate: [AppAuthGuard],
-    data: {breadcrumb: 'Add quiz', roles: ['Admin', 'EditQuizzes']}
+    data: {breadcrumb: 'Add quiz', roles: forAppCode(APP_CODE, ['Admin', 'EditQuizzes'])}
   },
   {
     path: 'quizzes/:quizId',
     component: QuizzesQuizComponent,
     canActivate: [AppAuthGuard],
-    data: {breadcrumb: 'Quiz', roles: ['Admin', 'EditQuizzes']},
+    data: {breadcrumb: 'Quiz', roles: forAppCode(APP_CODE, ['Admin', 'EditQuizzes'])},
     children: [
       {
         path: '',
