@@ -9,29 +9,30 @@ import {AppConfig, ServiceConfig, WorldskillsAngularLibModule, WSHttpConfig, WsH
 import {QuizzesComponent} from './quizzes/quizzes.component';
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {FontAwesomeModule} from '@fortawesome/angular-fontawesome';
-import {QuizzesQuizComponent} from './quizzes/quizzes-quiz/quizzes-quiz.component';
-import {QuizzesQuizFormComponent} from './quizzes/quizzes-quiz-form/quizzes-quiz-form.component';
+import {QuizzesQuizComponent} from './quizzes-quiz/quizzes-quiz.component';
+import {QuizzesQuizFormComponent} from './quizzes-quiz-form/quizzes-quiz-form.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {QuizzesQuizQuestionsComponent} from './quizzes/quizzes-quiz/quizzes-quiz-questions/quizzes-quiz-questions.component';
-import {QuizzesQuizPreviewComponent} from './quizzes/quizzes-quiz/quizzes-quiz-preview/quizzes-quiz-preview.component';
-import {QuizzesQuizTranslationsComponent} from './quizzes/quizzes-quiz/quizzes-quiz-translations/quizzes-quiz-translations.component';
-import {QuizzesQuizAttemptsComponent} from './quizzes/quizzes-quiz/quizzes-quiz-attempts/quizzes-quiz-attempts.component';
-import {QuizzesAttemptComponent} from './quizzes/quizzes-attempt/quizzes-attempt.component';
-import {QuizzesQuestionUpdateComponent} from './quizzes/quizzes-question-update/quizzes-question-update.component';
-import {QuizzesQuestionFormComponent} from './quizzes/quizzes-question-form/quizzes-question-form.component';
-import {QuizzesTranslationUpdateComponent} from './quizzes/quizzes-translation-update/quizzes-translation-update.component';
-import {QuizzesTranslationFormComponent} from './quizzes/quizzes-translation-form/quizzes-translation-form.component';
+import {QuizzesQuizQuestionsComponent} from './quizzes-quiz-questions/quizzes-quiz-questions.component';
+import {QuizzesQuizPreviewComponent} from './quizzes-quiz-preview/quizzes-quiz-preview.component';
+import {QuizzesQuizTranslationsComponent} from './quizzes-quiz-translations/quizzes-quiz-translations.component';
+import {QuizzesQuizAttemptsComponent} from './quizzes-quiz-attempts/quizzes-quiz-attempts.component';
+import {QuizzesAttemptComponent} from './quizzes-attempt/quizzes-attempt.component';
+import {QuizzesQuestionUpdateComponent} from './quizzes-question-update/quizzes-question-update.component';
+import {QuizzesQuestionFormComponent} from './quizzes-question-form/quizzes-question-form.component';
+import {QuizzesTranslationUpdateComponent} from './quizzes-translation-update/quizzes-translation-update.component';
+import {QuizzesTranslationFormComponent} from './quizzes-translation-form/quizzes-translation-form.component';
 import {QuizComponent} from './quiz/quiz.component';
-import {QuizzesQuizUpdateComponent} from './quizzes/quizzes-quiz-update/quizzes-quiz-update.component';
-import {QuizzesQuizCreateComponent} from './quizzes/quizzes-quiz-create/quizzes-quiz-create.component';
-import {QuizzesQuestionCreateComponent} from './quizzes/quizzes-question-create/quizzes-question-create.component';
-import {QuizzesTranslationCreateComponent} from './quizzes/quizzes-translation-create/quizzes-translation-create.component';
+import {QuizzesQuizUpdateComponent} from './quizzes-quiz-update/quizzes-quiz-update.component';
+import {QuizzesQuizCreateComponent} from './quizzes-quiz-create/quizzes-quiz-create.component';
+import {QuizzesQuestionCreateComponent} from './quizzes-question-create/quizzes-question-create.component';
+import {QuizzesTranslationCreateComponent} from './quizzes-translation-create/quizzes-translation-create.component';
 import {HttpInterceptorService} from '../services/http-interceptor/http-interceptor.service';
-import {CkEditorDirective} from './directives/ck-editor/ck-editor.directive';
+import {CkEditorDirective} from '../directives/ck-editor/ck-editor.directive';
 import {WsSpinnerComponent} from './ws-spinner/ws-spinner.component';
 import {LangCodeToNamePipe} from '../pipes/lang-code-to-name/lang-code-to-name.pipe';
 import {HomeComponent} from './home/home.component';
 import {ErrorComponent} from './error/error.component';
+import {NgSelectModule} from "@ng-select/ng-select";
 
 @NgModule({
   declarations: [
@@ -87,7 +88,8 @@ import {ErrorComponent} from './error/error.component';
     NgbModule,
     FontAwesomeModule,
     FormsModule,
-    ReactiveFormsModule
+    ReactiveFormsModule,
+    NgSelectModule
   ],
   providers: [
     {provide: HTTP_INTERCEPTORS, useClass: WsHttpInterceptor, multi: true},
