@@ -1,6 +1,9 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { HomeComponent } from './home.component';
+import {HomeComponent} from './home.component';
+import {RouterTestingModule} from "@angular/router/testing";
+import {AuthService, WorldskillsAngularLibModule} from "@worldskills/worldskills-angular-lib";
+import {HttpClientTestingModule} from "@angular/common/http/testing";
 
 describe('HomeComponent', () => {
   let component: HomeComponent;
@@ -8,9 +11,13 @@ describe('HomeComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ HomeComponent ]
+      imports: [RouterTestingModule, WorldskillsAngularLibModule, HttpClientTestingModule],
+      declarations: [HomeComponent],
+      providers: [
+        {provide: AuthService, useValue: {redirectOrReturn: () => undefined}},
+      ]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
