@@ -1,5 +1,4 @@
 import {Link, List} from './common';
-import {Entity} from './entity';
 import {Event} from './event';
 import {Skill} from './skill';
 import {I18nModel} from '@worldskills/worldskills-angular-lib';
@@ -12,7 +11,7 @@ export interface QuizRequest {
   reveal_correct_answers: boolean;
   url_learning: string;
   url_success: string;
-  ws_entity: Entity;
+  ws_entity: { id: number };
   event: Event;
   skill: Skill;
   active: boolean;
